requirements.txt
cbbpy
requests
beautifulsoup4
pandas
pyyaml